gcp_/16-01-2023/requirements.txt
Flask==3.0.0
requests
google-cloud-firestore
google-cloud-pubsub